Make travis happy
